iterlist: split immutable interface from mutable interface
New base class 'CachedIterator' is the basic interface for consuming an
iterable as it is needed. The underlying cache is a list, but it is not exposed
in this interface and the object actually behaves a lot like a tuple. The main
difference is that a CachedIterator will perform item-by-item equality and
comparison checking for any type of sequence.

IterTuple extends CachedIterator and can be used for tuple-like semantics: it will only
perform equlity/comparison checking against tuple and IterTuple instances.

IterList extends CachedIterator and retains the existing interface for mutable
operations, and additionally equality and comparisons are restricted to list
and IterList instances.